+++ research/object_detection/metrics/calibration_metrics.py
@@ -98,13 +98,16 @@ def expected_calibration_error(y_true, y_pred, nbins=20):
 
   with tf.control_dependencies([bin_ids]):
     update_bin_counts_op = tf.assign_add(
-        bin_counts, tf.to_float(tf.bincount(bin_ids, minlength=nbins)))
+        bin_counts, tf.cast(tf.bincount(bin_ids, minlength=nbins),
+                            dtype=tf.float32))
     update_bin_true_sum_op = tf.assign_add(
         bin_true_sum,
-        tf.to_float(tf.bincount(bin_ids, weights=y_true, minlength=nbins)))
+        tf.cast(tf.bincount(bin_ids, weights=y_true, minlength=nbins),
+                dtype=tf.float32))
     update_bin_preds_sum_op = tf.assign_add(
         bin_preds_sum,
-        tf.to_float(tf.bincount(bin_ids, weights=y_pred, minlength=nbins)))
+        tf.cast(tf.bincount(bin_ids, weights=y_pred, minlength=nbins),
+                dtype=tf.float32))
 
   ece_update_op = _ece_from_bins(
       update_bin_counts_op,
